% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_admfrom_ihunits}
\alias{check_admfrom_ihunits}
\title{Check 6
The time from injury to hospital admission should be within 24 hours
if the patient is coming from the Accident/Injury Site}
\usage{
check_admfrom_ihunits(form)
}
\arguments{
\item{form}{a dataframe containing ptstatus and form4.1}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check 6
The time from injury to hospital admission should be within 24 hours
if the patient is coming from the Accident/Injury Site
}
