% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_admission_followup}
\alias{check_admission_followup}
\title{Check that the follow-up occurs in-hospital if the patient is still 
in the hospital after 30 days}
\usage{
check_admission_followup(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1 and form 6.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that the follow-up occurs in-hospital if the patient is still 
in the hospital after 30 days
}
