% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_compno_7.x}
\alias{check_compno_7.x}
\title{Check 21
Check number of reported complications at follow-up is consistent with number of completed complication forms}
\usage{
check_compno_7.x(form)
}
\arguments{
\item{form}{a dataframe containing ptstatus, form6.1, and form 7.x}
}
\value{
a dataframe containing problematic entries
}
\description{
Check 21
Check number of reported complications at follow-up is consistent with number of completed complication forms
}
