% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_condate_hdcdate}
\alias{check_condate_hdcdate}
\title{Check that consent date is before discharge date}
\usage{
check_condate_hdcdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1 and form6.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that consent date is before discharge date
}
