% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_condate_hdcdate_dthdate}
\alias{check_condate_hdcdate_dthdate}
\title{Check that consent date is before or on discharge date or date of death}
\usage{
check_condate_hdcdate_dthdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1 and form6.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that consent date is before or on discharge date or date of death
}
