% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_condate_hspdate}
\alias{check_condate_hspdate}
\title{Check that consent date should be 0 - 30 days after hospital admission date}
\usage{
check_condate_hspdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1 and form4.1}
}
\value{
a dataframe containing problematic entries
}
\description{
condate after hspdate 0-30 days
}
