% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_condate_injdate}
\alias{check_condate_injdate}
\title{Check that consent date should be on the same day, or after the date of injury}
\usage{
check_condate_injdate(form1.1, form3.1)
}
\arguments{
\item{form1.1}{a dataframe containing form1.1}

\item{form3.1}{a dataframe containing form3.1}
}
\value{
a dataframe containing problematic entries
}
\description{
condate after injdate less than 92 days
}
