% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_dxdate_hspdate}
\alias{check_dxdate_hspdate}
\title{Check 22
Check that any omplication reported is within 30 days of hospital admission}
\usage{
check_dxdate_hspdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1, form4.1, form7.1, form7.2, form7.3, form7.4}
}
\value{
a dataframe containing problematic entries
}
\description{
Check 22
Check that any omplication reported is within 30 days of hospital admission
}
