% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2.1.R
\name{check_form2.1_textfield}
\alias{check_form2.1_textfield}
\title{Check for missing/extraneous textfields in form2.1}
\usage{
check_form2.1_textfield(form2.1)
}
\arguments{
\item{form}{form2.1}
}
\value{
problematic rows of form2.1
}
\description{
othoccup and p2q5 should be NA iff occup is not 13 (other)
}
