% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2.2.R
\name{check_form2.2_box10}
\alias{check_form2.2_box10}
\title{Check for missing/extraneous coding boxes in form2.2}
\usage{
check_form2.2_box10(form2.2)
}
\arguments{
\item{form}{form2.2}
}
\value{
problematic rows of form2.2
}
\description{
all p3q10x and comorb should be NA iff othercm is not 1 (checked)
}
