% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form3.2_box5}
\alias{check_form3.2_box5}
\title{Filters out invalid rows for box 5 form 3.2}
\usage{
check_form3.2_box5(form)
}
\arguments{
\item{form}{form3.2}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 5 form 3.2
}
