% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form3.2_box6}
\alias{check_form3.2_box6}
\title{Filters out invalid rows for box 6 of form3.2}
\usage{
check_form3.2_box6(form)
}
\arguments{
\item{form}{a form containing ptstatus and form3.2}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 6 of form3.2
}
