% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form3.2_box7_abdo}
\alias{check_form3.2_box7_abdo}
\title{Filters out invalid rows for the abdomen box 7 of form3.2}
\usage{
check_form3.2_box7_abdo(form)
}
\arguments{
\item{form}{form3.2}
}
\value{
a form containing ptstatus and a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for the abdomen box 7 of form3.2
}
