% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form3.2_box7_chest}
\alias{check_form3.2_box7_chest}
\title{Filters out invalid rows for the chest box 7 of form3.2}
\usage{
check_form3.2_box7_chest(form)
}
\arguments{
\item{form}{a form containing ptstatus and form3.2}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for the chest box 7 of form3.2
}
