% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form3.3_p50dist}
\alias{check_form3.3_p50dist}
\title{Filters out invalid rows for district code for form3.3}
\usage{
check_form3.3_p50dist(form)
}
\arguments{
\item{form}{a form containing ptstatus and form3.3}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for district code for form3.3
}
