% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form5.14_box1b}
\alias{check_form5.14_box1b}
\title{Filters out invalid rows for box 1b "Other time of application" of form5.14}
\usage{
check_form5.14_box1b(form)
}
\arguments{
\item{form}{a form containing ptstatus and form5.14}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 1b "Other time of application" of form5.14
}
