% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form5.14_box3}
\alias{check_form5.14_box3}
\title{Filters out invalid rows for box 3 of form5.14}
\usage{
check_form5.14_box3(form)
}
\arguments{
\item{form}{a form containing ptstatus and form5.14}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 3 of form5.14
}
