% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form5.1x_box1}
\alias{check_form5.1x_box1}
\title{Filters out invalid rows for box 1 of form 5.1}
\usage{
check_form5.1x_box1(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form5.1}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 1 of form 5.1
}
