% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form5.2x_box3a}
\alias{check_form5.2x_box3a}
\title{Filters out invalid rows for box 3 "other locations of treatment" of form5.2/6/10}
\usage{
check_form5.2x_box3a(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form5.2/6/10}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 3 "other locations of treatment" of form5.2/6/10
}
