% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form5.2x_box4}
\alias{check_form5.2x_box4}
\title{Filters out invalid rows for box 4 of form5.2/6/10}
\usage{
check_form5.2x_box4(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form5.2/6/10}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 4 of form5.2/6/10
}
