% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form5.3x_box5b}
\alias{check_form5.3x_box5b}
\title{Filters out invalid rows for box 5 "other non-operative treatment" of form5.3/7/11}
\usage{
check_form5.3x_box5b(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form5.3/7/11}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 5 "other non-operative treatment" of form5.3/7/11
}
