% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form5.4x_box9}
\alias{check_form5.4x_box9}
\title{Filters out invalid rows for box 9 of form5.4/8/12}
\usage{
check_form5.4x_box9(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form5.4/8/12}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 9 of form5.4/8/12
}
