% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form6.1_box1a}
\alias{check_form6.1_box1a}
\title{Filters out invalid rows for box 1 "Other patient discharge" of form6.1}
\usage{
check_form6.1_box1a(form)
}
\arguments{
\item{form}{a form containing ptstatus and form6.1}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 1 "Other patient discharge" of form6.1
}
