% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form6.1_box1b}
\alias{check_form6.1_box1b}
\title{Filters out invalid rows for box 1 "Cause of death" of form6.1}
\usage{
check_form6.1_box1b(form)
}
\arguments{
\item{form}{a form containing ptstatus and form6.1}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 1 "Cause of death" of form6.1
}
