% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form6.1_box2b}
\alias{check_form6.1_box2b}
\title{Filters out invalid rows for box 2 "Other cardiac complication" of form6.1}
\usage{
check_form6.1_box2b(form)
}
\arguments{
\item{form}{a form containing ptstatus and form6.1}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 2 "Other cardiac complication" of form6.1
}
