% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form7.x_box4c}
\alias{check_form7.x_box4c}
\title{Filters out invalid rows for box 4 "Other details" of form7.x}
\usage{
check_form7.x_box4c(form, rep)
}
\arguments{
\item{form}{a form containing ptstatus and form7.x}

\item{rep}{the repetition/iteration of the current form}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 4 "Other details" of form7.x
}
