% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFunctions.R
\name{check_form_box}
\alias{check_form_box}
\title{Filters out invalid rows in the form}
\usage{
check_form_box(form, other_val, other, text_box, coding)
}
\arguments{
\item{form}{any form}

\item{other_val}{int value corresponding to the 'other' option}

\item{other}{string name of the 'other' column}

\item{text_box}{string name of the 'text_box' column}

\item{coding}{a vector of column names for coding boxes}
}
\value{
a dataframe containing all the invalid rows
}
\description{
Filters out invalid rows in the form
}
