% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_form_box_II}
\alias{check_form_box_II}
\title{Filters out invalid rows in the form (Type II)}
\usage{
check_form_box_II(form, other_val, other, text_box, coding, none, normal)
}
\arguments{
\item{form}{any form}

\item{other_val}{int value corresponding to the 'other' option}

\item{other}{string name of the 'other' column}

\item{text_box}{string name of the 'text_box' column}

\item{coding}{a vector of column names for coding boxes}

\item{none}{column name of the 'none' check box}

\item{normal}{any normal check boxes}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows in the form (Type II)
}
