% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_form_box_III}
\alias{check_form_box_III}
\title{Filters out invalid rows in the form (Type III)}
\usage{
check_form_box_III(form, groups, types)
}
\arguments{
\item{form}{any form}

\item{groups}{each group contains a column name for the header check box
and the column names of its content elements}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows in the form (Type III)
}
