% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_formdate_ocrecyes}
\alias{check_formdate_ocrecyes}
\title{Check 16
Check that date of follow-up is 30 or more days after hospital admission}
\usage{
check_formdate_ocrecyes(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form4.1, and form6.1}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check 16
Check that date of follow-up is 30 or more days after hospital admission
}
