% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fracwith}
\alias{check_fracwith}
\title{Check only one fracture, if any, is checked per form5.1x}
\usage{
check_fracwith(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus and one set of form5.1x}

\item{rep}{which set of form5.x is checked}
}
\value{
a dataframe containing problematic entries
}
\description{
Check only one fracture, if any, is checked per form5.1x
}
