% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2.1.R
\name{check_from2.1_box5}
\alias{check_from2.1_box5}
\title{Filters out invalid rows for box 5 form 2.1}
\usage{
check_from2.1_box5(form)
}
\arguments{
\item{form}{form2.1}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 5 form 2.1
}
