% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_hspdate_injdate}
\alias{check_hspdate_injdate}
\title{Check that hospital admission date should be on the same day,
or after the date of injury}
\usage{
check_hspdate_injdate(form3.1, form4.1)
}
\arguments{
\item{form3.1}{a dataframe containing form3.1}

\item{form4.1}{a dataframe containing form4.1}
}
\value{
a dataframe containing problematic entries
}
\description{
hpsdate after injdate less than 92 days
}
