% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_iahrs_ihhrs}
\alias{check_iahrs_ihhrs}
\title{Time from injury to abx administration must be within the time interval (5h) from injury to hospital admission}
\usage{
check_iahrs_ihhrs(forms)
}
\arguments{
\item{merged_form}{a dataframe containing all forms}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Time from injury to abx administration must be within the time interval (5h) from injury to hospital admission
}
