% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ihhrs_ishrs_iahrs_12hrsafter}
\alias{check_ihhrs_ishrs_iahrs_12hrsafter}
\title{Check that the time from injury to abx administration should be no more than
 12 hours from the time from injury to def. stabilization}
\usage{
check_ihhrs_ishrs_iahrs_12hrsafter(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1, and form5.3}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the time from injury to abx administration should be no more than
 12 hours from the time from injury to def. stabilization
}
