% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ihhrs_ishrs_iahrs_1hrwithin}
\alias{check_ihhrs_ishrs_iahrs_1hrwithin}
\title{Check that the time from injury to def. stabilization should be within
 1 hour of the time from injury to abx administration}
\usage{
check_ihhrs_ishrs_iahrs_1hrwithin(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1, and form5.3}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the time from injury to def. stabilization should be within
 1 hour of the time from injury to abx administration
}
