% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_ihunits_ipunits}
\alias{check_ihunits_ipunits}
\title{Check that he time difference between the time from injury to hsp admission and 
the time from injury to prep solution in ER is within +/- 24 hr range}
\usage{
check_ihunits_ipunits(form)
}
\arguments{
\item{form}{a dataframe containing form1.1, form4.1, form5.14}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that he time difference between the time from injury to hsp admission and 
the time from injury to prep solution in ER is within +/- 24 hr range
}
