% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_initials}
\alias{check_initials}
\title{Check that initials between forms are consistent}
\usage{
check_initials(list)
}
\arguments{
\item{list}{list of forms to check for intial consistencies}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that initials between forms are consistent
}
