% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_injwith_frac}
\alias{check_injwith_frac}
\title{Check if the number of fractures, if any, is valid for the set form5.x
Also considers the logic of related selections for this check}
\usage{
check_injwith_frac(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus and one set of form5.x}

\item{rep}{which set of form5.x is checked}
}
\value{
a dataframe containing problematic entries
}
\description{
Check if the number of fractures, if any, is valid for the set form5.x
Also considers the logic of related selections for this check
}
