% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFunctions.R
\name{check_invalid}
\alias{check_invalid}
\title{Determines whether a row is valid or not}
\usage{
check_invalid(form, other_val, other, text_box, coding)
}
\arguments{
\item{form}{any form}

\item{other_val}{int value corresponding to the 'other' option}

\item{other}{string name of the 'other' column}

\item{text_box}{string name of the 'text_box' column}

\item{coding}{a vector of column names for coding boxes}
}
\value{
TRUE if all coding boxes are NA, otherwise FALSE
}
\description{
Determines whether a row is valid or not
}
