% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_form1.1}
\alias{check_invalid_form1.1}
\title{Check that all entries in form1.1 are filled with valid values}
\usage{
check_invalid_form1.1(form)
}
\arguments{
\item{form}{dataframe containing form1.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form1.1 are filled with valid values
}
