% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalids.R
\name{check_invalid_form2.1}
\alias{check_invalid_form2.1}
\title{Check that all entries in form2.1 are filled with valid values}
\usage{
check_invalid_form2.1(form)
}
\arguments{
\item{form}{dataframe containing form2.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form2.1 are filled with valid values
}
