% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_form2.2}
\alias{check_invalid_form2.2}
\title{Check that all entries in form2.2 are filled with valid values}
\usage{
check_invalid_form2.2(form)
}
\arguments{
\item{form}{dataframe containing form1.1 and form2.2}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form2.2 are filled with valid values
}
