% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_form3.1}
\alias{check_invalid_form3.1}
\title{Check that all entries in form3.1 are filled with valid values}
\usage{
check_invalid_form3.1(form, checkIntent = FALSE)
}
\arguments{
\item{form}{dataframe containing form1.1 and form3.1}

\item{checkIntent}{logical True to check question 4.1 Intent of Injury}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form3.1 are filled with valid values
}
