% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_form3.2_8}
\alias{check_invalid_form3.2_8}
\title{Check that all entries in Q8 form3.2 are filled with valid values}
\usage{
check_invalid_form3.2_8(form)
}
\arguments{
\item{form}{dataframe containing form3.2}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in Q8 form3.2 are filled with valid values
}
