% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalids.R
\name{check_invalid_form5.1x}
\alias{check_invalid_form5.1x}
\title{Check that all entries in form5.1x are filled with valid values}
\usage{
check_invalid_form5.1x(form, rep)
}
\arguments{
\item{form}{dataframe containing form1.1, form3.2, and form5.1x}

\item{rep}{which set of form}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form5.1x are filled with valid values
}
