% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_form5.2}
\alias{check_invalid_form5.2}
\title{Check that all entries in form5.2 are filled with valid values}
\usage{
check_invalid_form5.2(form)
}
\arguments{
\item{form}{dataframe containing form5.2}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that all entries in form5.2 are filled with valid values
}
