% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_I}
\alias{check_invalid_I}
\title{Determines whether a row is valid or not (Type I)}
\usage{
check_invalid_I(form, other_val, other, text_box, coding)
}
\arguments{
\item{form}{any form}

\item{other_val}{int value corresponding to the 'other' option}

\item{other}{string name of the 'other' column}

\item{text_box}{string name of the 'text_box' column}

\item{coding}{a vector of column names for coding boxes}
}
\value{
TRUE if a Type I invalid case has been found, otherwise returns FALSE
}
\description{
Determines whether a row is valid or not (Type I)
}
