% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_II}
\alias{check_invalid_II}
\title{Determines whether a row is valid or not (Type II)}
\usage{
check_invalid_II(form, other_val, other, text_box, coding, none, normal)
}
\arguments{
\item{form}{any form}

\item{other_val}{int value corresponding to the 'other' option}

\item{other}{string name of the 'other' column}

\item{text_box}{string name of the 'text_box' column}

\item{coding}{a vector of column names for coding boxes}

\item{none}{column name of the 'none' check box}

\item{normal}{any normal check boxes}
}
\value{
TRUE for a Type II invalid case, otherwise returns FALSE
}
\description{
Determines whether a row is valid or not (Type II)
}
