% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_III}
\alias{check_invalid_III}
\title{Determines whether a row is valid or not (Type III)}
\usage{
check_invalid_III(form, groups)
}
\arguments{
\item{form}{any form}

\item{groups}{each group contains a column name for the header check box
and the column names of its content elements}
}
\value{
TRUE for a Type III invalid case, otherwise returns FALSE
}
\description{
Determines whether a row is valid or not (Type III)
}
