% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_locabx}
\alias{check_locabx}
\title{Check that the time from injury to stabilization should be within 1 hour of
the time from injury to antibiotics administration if administered prior to
surgery}
\usage{
check_locabx(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1, and form5.3}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the time from injury to stabilization should be within 1 hour of
the time from injury to antibiotics administration if administered prior to
surgery
}
