% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_locabx_injurySite}
\alias{check_locabx_injurySite}
\title{Checks that if antibiotics was first administered at the injury scene, time
from injury to antibiotics administration is at most 5 hours before time from
injury to hospital admission}
\usage{
check_locabx_injurySite(form)
}
\arguments{
\item{form}{a form containing ptstatus and form 4.1}
}
\value{
a data frame containing all the invalid rows
}
\description{
Checks that if antibiotics was first administered at the injury scene, time
from injury to antibiotics administration is at most 5 hours before time from
injury to hospital admission
}
