% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_locabx_operatively}
\alias{check_locabx_operatively}
\title{Check that the time from injury to stabilization should be no more than 12
hours from the time from injury to antibiotics administration if administred
operatively}
\usage{
check_locabx_operatively(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1, and form5.3}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the time from injury to stabilization should be no more than 12
hours from the time from injury to antibiotics administration if administred
operatively
}
