% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_northinj}
\alias{check_northinj}
\title{Check that the number of ortho injuries stated on form 3.2 is consistent 
with the number of sets of injury forms completed}
\usage{
check_northinj(form)
}
\arguments{
\item{form}{dataframe containing all forms}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that the number of ortho injuries stated on form 3.2 is consistent 
with the number of sets of injury forms completed
}
