% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_northinj_form5.14}
\alias{check_northinj_form5.14}
\title{Check that the number of orthopedic injuries stated on form 3.2 is consistent 
with the Wound & Skin Prep form 5.14}
\usage{
check_northinj_form5.14(form)
}
\arguments{
\item{form}{a dataframe containing pststatus, form 3.2, and form 5.14}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that the number of orthopedic injuries stated on form 3.2 is consistent 
with the Wound & Skin Prep form 5.14
}
