% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_northinj_form5.x}
\alias{check_northinj_form5.x}
\title{Check that the number of orthopedic injuries stated on form 3.2 is consistent 
with the number of sets of injury forms completed}
\usage{
check_northinj_form5.x(form)
}
\arguments{
\item{form}{dataframe containing pststatus, form 3.2, and form5.x}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that the number of orthopedic injuries stated on form 3.2 is consistent 
with the number of sets of injury forms completed
}
