% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_openclos_NA}
\alias{check_openclos_NA}
\title{Check 13
Check that closed fracture injuries have have NA selected in form5.14}
\usage{
check_openclos_NA(form)
}
\arguments{
\item{form}{a dataframe containing form1.1, form5.1x, form5.14}
}
\value{
a dataframe containing problematic entries
}
\description{
Check 13
Check that closed fracture injuries have have NA selected in form5.14
}
