% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_openclose_iandd}
\alias{check_openclose_iandd}
\title{Check that the response to I&D is related to whether the fracture is open or closed 
in one set of form5.x are related}
\usage{
check_openclose_iandd(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus and one set of form5.x}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the response to I&D is related to whether the fracture is open or closed 
in one set of form5.x are related
}
