% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_operat_failsurg_delsurg}
\alias{check_operat_failsurg_delsurg}
\title{Check 12
Check that details on the patient's surgery is consistent}
\usage{
check_operat_failsurg_delsurg(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus, form3.2, form5.3, and form5.4}

\item{rep}{which set of form 5.3x}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check 12
Check that details on the patient's surgery is consistent
}
