% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_surgery}
\alias{check_surgery}
\title{Check that Had No Surgery on form 5.14 must be selected if patient did not receive srugery}
\usage{
check_surgery(form)
}
\arguments{
\item{form}{dataframe containing form5.2, form 5.3 and form 5.14}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that Had No Surgery on form 5.14 must be selected if patient did not receive srugery
}
