% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_surgery_consistency}
\alias{check_surgery_consistency}
\title{Check that details on the patient's surgery is consistent}
\usage{
check_surgery_consistency(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus, form3.2, form5.3, and form5.4}

\item{rep}{which set of form}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that details on the patient's surgery is consistent
}
