% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_treatment_method}
\alias{check_treatment_method}
\title{Check that the temporary form of stabilization (question 4, form 5.2) and the definitive form of stabilization (question 5, form 5.3) cannot be the same treatment, if same treatment, time must be different}
\usage{
check_treatment_method(form)
}
\arguments{
\item{form}{dataframe containing form5.2 and form 5.3}
}
\value{
a dataframe containing problematic entries
}
\description{
Check that the temporary form of stabilization (question 4, form 5.2) and the definitive form of stabilization (question 5, form 5.3) cannot be the same treatment, if same treatment, time must be different
}
