% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{count_all}
\alias{count_all}
\title{Counts all the items that return TRUE by the comparator}
\usage{
count_all(form, values, comp)
}
\arguments{
\item{form}{any form}

\item{values}{column names for values to be compared}

\item{comp}{comparator function}
}
\value{
total number of values that satisfy the comp conditions
}
\description{
Counts all the items that return TRUE by the comparator
}
