% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{diff_date}
\alias{diff_date}
\title{Parse dates and find their difference in days}
\usage{
diff_date(date1, date2)
}
\arguments{
\item{date1}{first character}

\item{date2}{second character date}
}
\value{
number of days between those dates or NA
}
\description{
date1 - date2 in days
returns NA if either date1 or date2 cannot be parsed
}
