% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_hip}
\alias{filter_hip}
\title{Filter out patients who have a hip injury (lpfem or rpfem)}
\usage{
filter_hip(form)
}
\arguments{
\item{form}{merged fpr,}
}
\value{
filtered tibble
}
\description{
Filter out patients who have a hip injury (lpfem or rpfem)
}
