% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_precovid}
\alias{filter_precovid}
\title{Filter out patients who were admitted to a hospital pre-covid
Defined as before March 11, 2020}
\usage{
filter_precovid(form)
}
\arguments{
\item{form}{form containing hspdate}
}
\value{
filtered tibble
}
\description{
Filter out patients who were admitted to a hospital pre-covid
Defined as before March 11, 2020
}
