% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{gt_zero}
\alias{gt_zero}
\title{Determines if val is positive}
\usage{
gt_zero(val)
}
\arguments{
\item{val}{numeric value or NA}
}
\value{
TRUE if val is greater than zero
}
\description{
Determines if val is positive
}
