% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{is_all}
\alias{is_all}
\title{Determines if all values produces the same output when passed
to a comparator}
\usage{
is_all(form, values, comp = is.na)
}
\arguments{
\item{form}{any form}

\item{values}{column names for values to be compared}

\item{comp}{comparator function (is.na by default unless otherwise specified)}
}
\value{
TRUE if all comparisons pass, otherwise returns FALSE
}
\description{
Determines if all values produces the same output when passed
to a comparator
}
