% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_invalid}
\alias{is_invalid}
\title{Check if numeric or character vector is invalid}
\usage{
is_invalid(x)
}
\arguments{
\item{x}{number or character}
}
\value{
TRUE/FALSE
}
\description{
invalid means 'M', or '*' for character or
<= -2 for numeric
}
