% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_invalid_na_or_n}
\alias{is_invalid_na_or_n}
\title{Check if numeric or character vector is invalid, NA or N}
\usage{
is_invalid_na_or_n(x)
}
\arguments{
\item{x}{number or character}
}
\value{
TRUE/FALSE
}
\description{
Check is_invalid and is_n for definitions of invalid and N
}
