% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_invalid_or_na}
\alias{is_invalid_or_na}
\title{Check if numeric or character vector is invalid or NA}
\usage{
is_invalid_or_na(x)
}
\arguments{
\item{x}{number or character}
}
\value{
TRUE/FALSE
}
\description{
Check is_invalid for definition of invalid
}
