% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_n}
\alias{is_n}
\title{Check if numeric or character vector is N}
\usage{
is_n(x)
}
\arguments{
\item{x}{number or character}
}
\value{
TRUE/FALSE
}
\description{
N means 'N' for character or
or -1 for numeric
}
