% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_na_coding}
\alias{is_na_coding}
\title{Determines if all coding boxes in a row have NA values}
\usage{
is_na_coding(form, coding)
}
\arguments{
\item{form}{any form}

\item{coding}{a vector of column names for coding boxes}
}
\value{
TRUE if all coding boxes are NA, otherwise FALSE
}
\description{
Determines if all coding boxes in a row have NA values
}
