% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_zero}
\alias{is_zero}
\title{Determines if val is zero}
\usage{
is_zero(val)
}
\arguments{
\item{val}{numeric value or NA}
}
\value{
TRUE if val is equal to zero
}
\description{
Determines if val is zero
}
