% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_forms}
\alias{merge_forms}
\title{Merge a list of forms}
\usage{
merge_forms(forms)
}
\arguments{
\item{forms}{list of parsed (not raw) forms}
}
\value{
merge dataframe
}
\description{
Merge a list of parsed forms by region, sit, studyid, and ptintit
}
