% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_dec}
\alias{parse_dec}
\title{Convert string to decimal number}
\usage{
parse_dec(x)
}
\arguments{
\item{x}{string}
}
\value{
decimal number
}
\description{
M will be converted to -1.0
* (lost) will be converted to -2.0
}
