% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_dmY}
\alias{parse_dmY}
\title{Convert string in d/m/Y format to dates}
\usage{
parse_dmY(dates)
}
\arguments{
\item{dates}{d/m/Y string}
}
\value{
date or NA if date is invalid
}
\description{
Check parse_date for details
}
