% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_int}
\alias{parse_int}
\title{Convert string to integer}
\usage{
parse_int(x)
}
\arguments{
\item{x}{string}
}
\value{
integer
}
\description{
M will be converted to -1
* (lost) will be converted to -2
}
