% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_num}
\alias{parse_num}
\title{Convert string to number}
\usage{
parse_num(x)
}
\arguments{
\item{x}{string}
}
\value{
integer
}
\description{
M will be converted to -2
* (lost) will be converted to -3
N will be converted to -1
}
