% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_region}
\alias{parse_region}
\title{Convert id to site region}
\usage{
parse_region(id)
}
\arguments{
\item{id}{studyid}
}
\value{
string of site region or NA
}
\description{
Convert studyid to site region
}
