% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_site}
\alias{parse_site}
\title{Convert id to site location}
\usage{
parse_site(id)
}
\arguments{
\item{id}{studyid}
}
\value{
string of site location or NA
}
\description{
Convert studyid to site location
}
