% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_site_c}
\alias{parse_site_c}
\title{Convert id to site location}
\usage{
parse_site_c(id)
}
\arguments{
\item{id}{studyid}
}
\value{
string of site location or NA
}
\description{
Convert studyid to site location
}
