% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{percent}
\alias{percent}
\title{Format a number to a percentage}
\usage{
percent(x)
}
\arguments{
\item{x}{number}
}
\value{
percentage as a string
}
\description{
the percentage will have 2 digits after the decimal
}
