% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_form1.1}
\alias{read_form1.1}
\title{Read in form1.1}
\usage{
read_form1.1(path)
}
\arguments{
\item{path}{path to file}
}
\value{
form1.1 dataframe
}
\description{
Read in form1.1 from path
Convert integers to logicals
}
