% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2.1.R
\name{read_form2.1}
\alias{read_form2.1}
\title{Read in form2.1}
\usage{
read_form2.1(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form2.1 dataframe if raw is FALSE
}
\description{
Read in form2.1 from file path
}
