% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form2.2.R
\name{read_form2.2}
\alias{read_form2.2}
\title{Read in form2.2}
\usage{
read_form2.2(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form2.2 dataframe
}
\description{
Read in form2.2 from file path
}
