% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form3.1.R
\name{read_form3.1}
\alias{read_form3.1}
\title{Read in form3.1}
\usage{
read_form3.1(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form3.1 dataframe
}
\description{
Read in form3.1 from file path
}
