% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form3.2.R
\name{read_form3.2}
\alias{read_form3.2}
\title{Read in form3.2}
\usage{
read_form3.2(path, raw = TRUE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form3.2 dataframe
}
\description{
Read in form3.2 from file path
}
