% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form3.3.R
\name{read_form3.3}
\alias{read_form3.3}
\title{Read in form3.3}
\usage{
read_form3.3(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form3.3 dataframe
}
\description{
Read in form3.3 from file path
}
