% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form4.1.R
\name{read_form4.1}
\alias{read_form4.1}
\title{Read in form4.1}
\usage{
read_form4.1(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form4.1 dataframe
}
\description{
Read in form4.1 from file path
}
