% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.10.R
\name{read_form5.10}
\alias{read_form5.10}
\title{Read in form5.10}
\usage{
read_form5.10(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.10 dataframe
}
\description{
Read in form5.10 from file path
}
