% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.11.R
\name{read_form5.11}
\alias{read_form5.11}
\title{Read in form5.11}
\usage{
read_form5.11(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.11 dataframe
}
\description{
Read in form5.11 from file path
}
