% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.12.R
\name{read_form5.12}
\alias{read_form5.12}
\title{Read in form5.12}
\usage{
read_form5.12(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.12 dataframe
}
\description{
Read in form5.12 from file path
}
