% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.13.R
\name{read_form5.13}
\alias{read_form5.13}
\title{Read in form5.13}
\usage{
read_form5.13(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.13 dataframe
}
\description{
Read in form5.13 from file path
}
