% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.14.R
\name{read_form5.14}
\alias{read_form5.14}
\title{Read in form5.14}
\usage{
read_form5.14(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.14 dataframe
}
\description{
Read in form5.14 from file path
}
