% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.1.R
\name{read_form5.159}
\alias{read_form5.159}
\title{Read in form5.1, 5.5, or 5.9}
\usage{
read_form5.159(path, raw, sfx)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}

\item{sfx}{what to append to column names}
}
\value{
form5.1, 5.5, or 5.9 dataframe
}
\description{
Read in form5.1, 5.5, or 5.9 from file path
}
