% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.2.R
\name{read_form5.2610}
\alias{read_form5.2610}
\title{Read in form5.2, 5.6, or 5.10}
\usage{
read_form5.2610(path, raw, sfx)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}

\item{sfx}{what to append to column names}
}
\value{
form5.2, 5.6, or 5.10 dataframe
}
\description{
Read in form5.2, 5.6, or 5.10 from file path
}
