% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.3.R
\name{read_form5.3}
\alias{read_form5.3}
\title{Read in form5.3}
\usage{
read_form5.3(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.3 dataframe
}
\description{
Read in form5.3 from file path
}
