% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.3.R
\name{read_form5.3711}
\alias{read_form5.3711}
\title{Read in form5.3, 5.7, or 5.11}
\usage{
read_form5.3711(path, raw, sfx)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}

\item{sfx}{what to append to column names}
}
\value{
form5.3, 5.7, or 5.11 dataframe
}
\description{
Read in form5.3, 5.7, or 5.11 from file path
}
