% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.4.R
\name{read_form5.4}
\alias{read_form5.4}
\title{Read in form5.4}
\usage{
read_form5.4(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.4 dataframe
}
\description{
Read in form5.4 from file path
}
