% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.4.R
\name{read_form5.4812}
\alias{read_form5.4812}
\title{Read in form5.4, 5.8, or 5.12}
\usage{
read_form5.4812(path, raw, sfx)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}

\item{sfx}{what to append to column names}
}
\value{
form5.4, 5.8, or 5.12 dataframe
}
\description{
Read in form5.4, 5.8, or 5.12 from file path
}
