% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.5.R
\name{read_form5.5}
\alias{read_form5.5}
\title{Read in form5.5}
\usage{
read_form5.5(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.5 dataframe
}
\description{
Read in form5.5 from file path
}
