% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.6.R
\name{read_form5.6}
\alias{read_form5.6}
\title{Read in form5.6}
\usage{
read_form5.6(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.6 dataframe
}
\description{
Read in form5.6 from file path
}
