% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.7.R
\name{read_form5.7}
\alias{read_form5.7}
\title{Read in form5.7}
\usage{
read_form5.7(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.7 dataframe
}
\description{
Read in form5.7 from file path
}
