% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form5.8.R
\name{read_form5.8}
\alias{read_form5.8}
\title{Read in form5.8}
\usage{
read_form5.8(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.8 dataframe
}
\description{
Read in form5.8 from file path
}
