% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form5.9.R
\name{read_form5.9}
\alias{read_form5.9}
\title{Read in form5.9}
\usage{
read_form5.9(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form5.9 dataframe
}
\description{
Read in form5.9 from file path
}
