% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form6.1.R
\name{read_form6.1}
\alias{read_form6.1}
\title{Read in form6.1}
\usage{
read_form6.1(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form6.1 dataframe
}
\description{
Read in form6.1 from file path
}
