% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form7.2.R
\name{read_form7.2}
\alias{read_form7.2}
\title{Read in form7.2}
\usage{
read_form7.2(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form7.2 dataframe
}
\description{
Read in form7.2 from file path
}
