% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form7.3.R
\name{read_form7.3}
\alias{read_form7.3}
\title{Read in form7.3}
\usage{
read_form7.3(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form7.3 dataframe
}
\description{
Read in form7.3 from file path
}
