% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form7.4.R
\name{read_form7.4}
\alias{read_form7.4}
\title{Read in form7.4}
\usage{
read_form7.4(path, raw = FALSE)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}
}
\value{
form7.4 dataframe
}
\description{
Read in form7.4 from file path
}
