% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_form7.1.R
\name{read_form7.x}
\alias{read_form7.x}
\title{Read in form7.x}
\usage{
read_form7.x(path, raw, sfx)
}
\arguments{
\item{path}{path to file}

\item{raw}{if TRUE, return raw data}

\item{sfx}{what to append to column names}
}
\value{
form7.x dataframe
}
\description{
Read in form7.x from file path
}
