% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_cat}
\alias{summarise_cat}
\title{Summarise a categorical column}
\usage{
summarise_cat(grouped_form, colname, vals)
}
\arguments{
\item{grouped_form}{any grouped dataframe}

\item{colname}{name to column to summarise}

\item{vals}{values in that column}
}
\value{
summary dataframe
}
\description{
Summarise the number and percentage of values in a column
}
