% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_catval}
\alias{summarise_catval}
\title{Summarise a value in a column}
\usage{
summarise_catval(grouped_form, colname, val)
}
\arguments{
\item{grouped_form}{any grouped dataframe}

\item{colname}{name to column to summarise}
}
\value{
summary dataframe
}
\description{
Summarise the number and percentage of that value in a column
}
