% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_cont}
\alias{summarise_cont}
\title{Summarise a continuous column}
\usage{
summarise_cont(grouped_form, colname)
}
\arguments{
\item{grouped_form}{any grouped dataframe}

\item{colname}{name to column to summarise}
}
\value{
summary dataframe
}
\description{
Summarise min, max, mean, and sd of values in that column
}
