% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_date}
\alias{summarise_date}
\title{Summarise a date column}
\usage{
summarise_date(grouped_form, colname)
}
\arguments{
\item{grouped_form}{any grouped dataframe}

\item{colname}{name to column to summarise}
}
\value{
summary dataframe
}
\description{
Summarise min, max of dates in that column
}
