% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_form}
\alias{summarise_form}
\title{Summarise a form}
\usage{
summarise_form(form, cols, ...)
}
\arguments{
\item{form}{any form}

\item{cols}{names of the columns to summarise}

\item{...}{names of the columns to group by}
}
\value{
summary dataframe
}
\description{
Summarise the given columns of a form
Note that cols and ... cannot contain the same columns
}
