% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarise_form1.1}
\alias{summarise_form1.1}
\title{Same as summarize_form1.1}
\usage{
summarise_form1.1(form)
}
\arguments{
\item{form}{dataframe containing all infor of form1.1}
}
\value{
dataframe, summary table
}
\description{
An alias  to summarize_form1.1
}
