% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_NA}
\alias{summarise_NA}
\title{Summarise NAs in a column}
\usage{
summarise_NA(grouped_form, colname)
}
\arguments{
\item{grouped_form}{any grouped dataframe}

\item{colname}{name to column to summarise}
}
\value{
summary dataframe
}
\description{
Summarise the number and percentage of NA in  acolumn
}
