% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarize_form}
\alias{summarize_form}
\title{Same as summarise_form}
\usage{
summarize_form(form, cols, ...)
}
\arguments{
\item{cols}{names of the columns to summarise}

\item{...}{names of the columns to group by}
}
\value{
summary dataframe
}
\description{
#' @param form any form
}
