% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_form1.1}
\alias{summarize_form1.1}
\title{Summarize form1.1}
\usage{
summarize_form1.1(form)
}
\arguments{
\item{form}{dataframe containing all infor of form1.1}
}
\value{
dataframe, summary table
}
\description{
Takes in any form containing all information of form1.1
For example, form1.1 merged with form2.1 will work
Returns a summary of the freqencies and percentages
of each datapoint
}
