% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_condate_injdate}
\alias{check_condate_injdate}
\title{Check 1
Check that consent date should be on the same day, or after the date of injury}
\usage{
check_condate_injdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1 and form3.1}
}
\value{
a dataframe containing problematic entries
}
\description{
condate after injdate 0-92 days
}
