% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boxes.R
\name{check_form2.2_box10}
\alias{check_form2.2_box10}
\title{Filters out invalid rows for box 10 of form2.2}
\usage{
check_form2.2_box10(form)
}
\arguments{
\item{form}{a form containing ptstatus and form2.2}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 10 of form2.2
}
