% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_form4.1_box2}
\alias{check_form4.1_box2}
\title{Filters out invalid rows for box 2 of form4.1}
\usage{
check_form4.1_box2(form)
}
\arguments{
\item{form}{a form containing ptstatus and form4.1}
}
\value{
a data frame containing all the invalid rows
}
\description{
Filters out invalid rows for box 2 of form4.1
}
