% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_fracwith_diswith}
\alias{check_fracwith_diswith}
\title{Check 8 and 9
Check that the location of fracture and the location of dislocation in one
set of form5.x are related}
\usage{
check_fracwith_diswith(form, rep)
}
\arguments{
\item{form}{dataframe containing ptstatus and one set of form5.x}

\item{rep}{which set of form5.x is checked}
}
\value{
a dataframe containing problematic entries
}
\description{
Check 8 and 9
Check that the location of fracture and the location of dislocation in one
set of form5.x are related
}
