% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_hspdate_injdate}
\alias{check_hspdate_injdate}
\title{Check that hospital admission date should be on the same day,
or after the date of injury}
\usage{
check_hspdate_injdate(merged_form)
}
\arguments{
\item{merged_form}{a dataframe containing form1.1, form3.1, and form4.1}
}
\value{
a dataframe containing problematic entries
}
\description{
hpsdate after injdate 0-92 days
}
