% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_injdate_hspdate}
\alias{check_injdate_hspdate}
\title{Check 5
Check that time from injury to hsp admission should be within +/- 24 hrs
range of difference between injdate and hspdate}
\usage{
check_injdate_hspdate(form)
}
\arguments{
\item{form}{a dataframe containing form1.1, form3.1, form4.1}
}
\value{
a dataframe containing problematic entries
}
\description{
Check 5
Check that time from injury to hsp admission should be within +/- 24 hrs
range of difference between injdate and hspdate
}
