% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_invalid_boxes}
\alias{check_invalid_boxes}
\title{Helper to filter out invalid coding boxes}
\usage{
check_invalid_boxes(
  form,
  other_vals,
  other_fields,
  text_fields,
  coding_boxes,
  comment
)
}
\arguments{
\item{form}{any form containing ptstatus, other_fields, text_fields, and coding_boxes}

\item{other_vals}{int value corresponding to the 'other' option}

\item{other_fields}{string name of the 'other_fields' column}

\item{text_fields}{string name of the 'text_fields' column}

\item{coding_boxes}{a vector of column names for coding boxes}

\item{comment}{comment about specific check}
}
\value{
a data frame containing all the invalid rows
}
\description{
Helper to filter out invalid coding boxes
}
