% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_checks.R
\name{check_locabx}
\alias{check_locabx}
\title{Check that the time from injury to the first antibiotic administration must
be consistent with the location of the first administration}
\usage{
check_locabx(form)
}
\arguments{
\item{form}{dataframe containing ptstatus, form 4.1, and form5.3}
}
\value{
a dataframe containing problematic entries with relevant columns
}
\description{
Check that the time from injury to the first antibiotic administration must
be consistent with the location of the first administration
}
