% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{percent}
\alias{percent}
\title{Convert decimal to percentage}
\usage{
percent(x)
}
\arguments{
\item{x}{decimal number}
}
\value{
string
}
\description{
Convert a decimal number to a percentage as string
with 2 decimal places
0.3 -> 30%
}
